requirements.txt
django==1.7.7
pillow==2.7.0
psycopg2==2.6
gunicorn==19.3.0
python-social-auth==0.2.3
django-debug-toolbar==1.3.0